search service: switch to protobuf

+++ services/search/ops.go
@@ -16,10 +16,10 @@ type ESClient struct {
 
 type Post struct {
 	Title, Body string
-	Id          uint
-	UserID      int
+	Id          uint64
+	UserID      uint64
 	LastUpdate  int64
-	Likes       int
+	Likes       int64
 }
 
 var (
@@ -83,23 +83,23 @@ func (esc *ESClient) Index(post Post) error {
 	return err
 }
 
-func (esc *ESClient) Search(term string, from, total int) ([]uint, error) {
+func (esc *ESClient) Search(term string, from, total uint64) ([]uint64, error) {
 	ctx := context.TODO()
 	query := elastic.NewMultiMatchQuery(term, "Title", "Body")
 	searchResult, err := esc.client.Search().
 		Index(esc.indexName).
 		Query(query).
 		SortBy(elastic.NewFieldSort("LastUpdate").Asc()).
-		From(from).Size(total).
+		From(int(from)).Size(int(total)).
 		Do(ctx)
 	if err != nil {
-		return []uint{}, err
+		return []uint64{}, err
 	}
-	results := []uint{}
+	results := []uint64{}
 	for _, hit := range searchResult.Hits.Hits {
 		var p Post
 		if err := json.Unmarshal(hit.Source, &p); err != nil {
-			return []uint{}, err
+			return []uint64{}, err
 		}
 		results = append(results, p.Id)
 	}
@@ -107,7 +107,7 @@ func (esc *ESClient) Search(term string, from, total int) ([]uint, error) {
 	return results, nil
 }
 
-func (esc *ESClient) UpdateLikes(id uint, likes int) error {
+func (esc *ESClient) UpdateLikes(id uint64, likes int64) error {
 	ctx := context.TODO()
 	query := elastic.NewTermQuery("Id", id)
 	script := elastic.NewScript(fmt.Sprintf("ctx._source.Likes = %v", likes))
@@ -119,7 +119,7 @@ func (esc *ESClient) UpdateLikes(id uint, likes int) error {
 	return err
 }
 
-func (esc *ESClient) Delete(id uint) error {
+func (esc *ESClient) Delete(id uint64) error {
 	ctx := context.TODO()
 	query := elastic.NewTermQuery("Id", id)
 	_, err := esc.client.DeleteByQuery().
@@ -129,7 +129,7 @@ func (esc *ESClient) Delete(id uint) error {
 	return err
 }
 
-func (esc *ESClient) UpdateLastUpdate(id uint, lastUpdate int64) error {
+func (esc *ESClient) UpdateLastUpdate(id uint64, lastUpdate int64) error {
 	ctx := context.TODO()
 	query := elastic.NewTermQuery("Id", id)
 	script := elastic.NewScript(fmt.Sprintf("ctx._source.LastUpdate = %v", lastUpdate))
